requirements.txt
arrow==0.15.5
asgiref==3.2.10
asn1crypto==1.3.0
attrdict==2.0.1
attrs==19.3.0
bip32utils==0.3.post4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==6.6
coincurve==13.0.0
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.10.1
Django==3.0.2
django-extensions==2.2.6
django-rest-framework==0.1.0
djangorestframework==3.11.0
docker-py==1.8.0
docopt==0.6.2
drf-yasg==1.17.0
ecdsa==0.15
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
flake8==3.7.9
funcsigs==1.0.2
future==0.18.2
gunicorn==20.0.4
hexbytes==0.2.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
ipfshttpclient==0.4.12
itypes==1.1.0
jedi==0.16.0
Jinja2==2.10.3
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
lru-dict==1.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
mnemonic==0.13
more-itertools==8.2.0
multiaddr==0.0.9
munch==2.5.0
mypy-extensions==0.4.3
netaddr==0.7.19
packaging==20.1
parsimonious==0.8.1
parso==0.6.0
path==13.1.0
path.py==12.4.0
pbkdf2==1.3
pexpect==4.8.0
pika==1.1.0
pluggy==0.13.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psycopg2-binary==2.8.4
ptpython==2.0.6
ptyprocess==0.6.0
py==1.8.1
py-ecc==2.0.0
pyaes==1.6.1
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.9.4
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
pysha3==1.0.2
pytest==5.3.5
python-bitcoinrpc==1.0
python-dateutil==2.8.1
python-jsonrpc==0.10.0
pytz==2019.3
pywallet==0.1.0
PyYAML==5.3
repoze.lru==0.7
requests==2.23.0
rlp==1.2.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
scrypt==0.8.13
sha256==0.1
six==1.14.0
sqlparse==0.3.0
tabulate==0.8.6
toolz==0.10.0
two1==3.10.9
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
varint==1.0.2
wcwidth==0.1.8
web3==5.6.0
websocket-client==0.57.0
websockets==8.1
zipp==3.1.0
Pillow==7.2.0
